feat: further improve folder structure
Also added utility functions for use without framework.
