+++ src/base/index.ts
@@ -1,0 +1,22 @@
+/*
+ * This program has been developed by students from the bachelor
+ * Computer Science at Utrecht University within the Software Project course.
+ *
+ * © Copyright Utrecht University
+ * (Department of Information and Computing Sciences)
+ */
+
+/**
+ * @packageDocumentation Contains the basic GraphPolaris visualisation plugin API. This is required for any plugin, regardless of the framework used.
+ *
+ * @categoryDescription Graph data
+ * These types are used for transferring raw graph data.
+ * @categoryDescription Message event listeners
+ * These are functions to listen for specific types of messages arriving.
+ */
+
+export * from "./eventListeners";
+export * from "./graphQueryResult.types";
+export * from "./machineLearning.types";
+export * from "./schema.types";
+export * from "./settings.types";
