+++ src/react/hooks.ts
@@ -6,8 +6,9 @@
  * (Department of Information and Computing Sciences)
  */
 
-import { useEffect, useMemo, useState, useContext, createContext } from "react";
-import { Settings, ReceiveMessage, SendMessage } from "./message";
+import { useEffect, useMemo, useState } from "react";
+import type { ReceiveMessage } from "../base/message.types";
+import { receiveMessage, sendMessage, type Settings } from "../base";
 
 /**
  * The context for providing the window object to the hooks and components.
@@ -90,29 +91,15 @@ function useMessage<
   const window = useContext(WindowContext);
   const [message, setMessage] = useState(defaultValue);
 
-  function updateMessage(e: MessageEvent<ReceiveMessage>) {
-    const data = e.data;
-
-    if (data.type === typeFilter) {
-      setMessage(data.data as TData);
-    }
-  }
+  const updateMessage = (data: TData) => setMessage(data);
 
   /* eslint-disable react-hooks/exhaustive-deps -- dependency cannot change */
   useEffect(() => {
-    window.addEventListener("message", updateMessage);
-    return () => window.removeEventListener("message", updateMessage);
+    return receiveMessage(typeFilter, updateMessage as any);
   }, []);
   /* eslint-enable react-hooks/exhaustive-deps */
 
   return useMemo(() => message, [message]);
-}
-
-/**
- * Sends a message to the parent window.
- */
-function sendMessage(message: SendMessage) {
-  window.top?.postMessage(message, "*");
 }
 
 /**
