+++ typedoc.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://typedoc.org/schema.json",
   "name": "@graphpolaris/vis-api",
-  "entryPoints": ["./src/index.ts", "./src/types/index.ts"],
+  "entryPoints": ["./src/react/index.ts", "./src/base/index.ts"],
   "categorizeByGroup": false,
   "gitRemote": "origin",
   "searchCategoryBoosts": {
